transactons updated

+++ src/components/Dashboard/Create_Courses.js
@@ -1,8 +1,11 @@
 import {
   Box,
+  Button,
   Card,
+  Divider,
   IconButton,
   makeStyles,
+  TextField,
   Typography,
 } from "@material-ui/core";
 import React, { useEffect, useState } from "react";
@@ -12,6 +15,7 @@ import AddCircleOutlineIcon from "@mui/icons-material/AddCircleOutline";
 import EditIcon from "@mui/icons-material/Edit";
 import DeleteIcon from "@mui/icons-material/Delete";
 import { useNavigate } from "react-router-dom";
+import { CSVLink } from "react-csv";
 import {
   collection,
   onSnapshot,
@@ -23,18 +27,48 @@ import {
   getDocs,
 } from "firebase/firestore";
 import { db } from "../../firebase";
-import DataTableExtensions from "react-data-table-component-extensions";
+
 import "react-data-table-component-extensions/dist/index.css";
 import { paths } from "../Routes/paths";
+import { toast, ToastContainer } from "react-toastify";
+import swal from "sweetalert";
 const useStyles = makeStyles(() => ({
   title: {
     zIndex: 999,
   },
+  csvFile: {
+    width: "15%",
+    textDecoration: "none",
+    display: "flex",
+    justifyContent: " space-around",
+    color: "#fff",
+    marginTop: "10px",
+    backgroundColor: "blue",
+    padding: "10px 0",
+    borderRadius: "5px",
+  },
+  csvFileParent: {
+    display: "flex",
+    justifyContent: "space-between",
+    width: "99%",
+  },
+  typo: {
+    marginLeft: "10px",
+  },
+  search: {
+    marginTop: "10px",
+  },
 }));
 const Create_Courses = () => {
+  function numberWithCommas(x) {
+    return x.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ",");
+  }
   const classes = useStyles();
   const navigate = useNavigate();
   const [data, setData] = useState([]);
+  const [FilteredData, setFilteredData] = useState([]);
+  const [open, setOpen] = useState(false);
+  const [q, setQ] = useState("");
   const actionEdit = async (id) => {
     try {
       const netEditElement = data.find((elemnt) => elemnt.id === id);
@@ -45,12 +79,30 @@ const Create_Courses = () => {
   };
 
   const actionDelete = async (id) => {
-    await deleteDoc(doc(db, "addCourse", id));
-    const oneIndex = data.filter((val) => val.id !== id);
-    const cloned = [...data];
-    cloned.splice(oneIndex, 1);
-    setData(cloned);
+    setOpen(true);
+    swal({
+      title: "Are you sure?",
+      text: "Once deleted, you will not be able to recover this imaginary file!",
+      icon: "warning",
+      buttons: true,
+      dangerMode: true,
+    }).then((willDelete) => {
+      if (willDelete) {
+        deleteDoc(doc(db, "addCourse", id));
+        const oneIndex = data.filter((val) => val.id == id);
+        const cloned = [...data];
+        cloned.splice(oneIndex, 1);
+        setData(cloned);
+        setOpen(false);
+        swal("Poof! Your imaginary file has been deleted!", {
+          icon: "success",
+        });
+      } else {
+        swal("Your imaginary file is safe!");
+      }
+    });
   };
+
   useEffect(() => {
     const q = query(collection(db, "addCourse"), orderBy("timestamp", "desc"));
     const unsubscribe = onSnapshot(q, (querySnapshot) => {
@@ -70,78 +122,99 @@ const Create_Courses = () => {
   const columns = [
     {
       name: "Course Id",
-      selector: (row) => row.id,
+      selector: (row) => row.s_id,
       sortable: true,
-      reorder: true,
+      reorder: false,
     },
     {
       name: "Class Code",
       selector: (row) => row.Class_Code,
-      reorder: true,
+      reorder: false,
     },
     {
       name: "Class Name",
       selector: (row) => row.class,
-      reorder: true,
+      reorder: false,
     },
 
     {
       name: "Admission Fee ",
-      selector: (row) => row.AdmissionFee,
-      reorder: true,
+      selector: (row) => " ₹ " + numberWithCommas(row.AdmissionFee.toFixed(2)),
+      reorder: false,
     },
     {
       name: "Per/Monthely Fee",
-      selector: (row) => row.monthelyFee,
-      reorder: true,
+      selector: (row) => " ₹ " + numberWithCommas(row.monthelyFee.toFixed(2)),
+      reorder: false,
     },
     {
       name: "Actions",
       selector: (row) => (
-        <>
+        <Box display="flex">
           <IconButton onClick={() => actionEdit(row?.id)}>
             <EditIcon />
           </IconButton>
           <IconButton onClick={() => actionDelete(row?.id)}>
             <DeleteIcon />
           </IconButton>
-        </>
+        </Box>
       ),
       ignoreRowClick: true,
       allowOverflow: true,
       button: true,
     },
   ];
-  const tableData = {
-    columns,
-    data,
+  const handleSearch = (e) => {
+    setQ(e.target.value);
   };
-
+  useEffect(() => {
+    const value = data.filter((item) =>
+      item.Class_Code.toLowerCase().includes(q)
+    );
+    setFilteredData(value);
+  }, [q, data]);
   return (
     <Box>
+      <ToastContainer />
       <Card>
-        <DataTableExtensions {...tableData} filterPlaceholder={"Search"}>
-          <DataTable
-            title={
-              <Box>
-                {"Courses"}
-                {
-                  <IconButton onClick={addHandler} color="primary">
-                    <AddCircleOutlineIcon />
-                  </IconButton>
-                }
-              </Box>
-            }
-            fixedHeader
-            columns={columns}
+        <div className={classes.csvFileParent}>
+          <Box display="flex" alignItems="center" className={classes.typo}>
+            <Typography variant="h5"> Courses </Typography>
+            <IconButton onClick={addHandler} color="primary">
+              <AddCircleOutlineIcon />
+            </IconButton>
+          </Box>
+
+          <CSVLink
             data={data}
-            defaultSortField="id"
-            defaultSortAsc={false}
-            striped
-            highlightOnHover
-            pagination
-          />
-        </DataTableExtensions>
+            filename="students.csv"
+            target="_blank"
+            className={classes.csvFile}
+          >
+            Export CSV
+          </CSVLink>
+        </div>
+
+        <DataTable
+          actions={
+            <Box className={classes.search}>
+              {/* <RangePicker onDateChanges={onDateChanges} /> */}
+              <TextField
+                label="Search"
+                variant="outlined"
+                value={q}
+                onChange={handleSearch}
+              />
+            </Box>
+          }
+          columns={columns}
+          data={FilteredData}
+          defaultSortField="id"
+          defaultSortAsc={false}
+          striped
+          highlightOnHover
+          pagination
+        />
       </Card>
     </Box>
   );
